+++ src/Bundles/Storage/StorageFactory.php
@@ -11,6 +11,7 @@ use PHPUnuhi\Bundles\Storage\PHP\PhpStorage;
 use PHPUnuhi\Bundles\Storage\PO\PoStorage;
 use PHPUnuhi\Bundles\Storage\RESX\ResxStorage;
 use PHPUnuhi\Bundles\Storage\Shopware6\Shopware6Storage;
+use PHPUnuhi\Bundles\Storage\Strings\StringsStorage;
 use PHPUnuhi\Bundles\Storage\YAML\YamlStorage;
 use PHPUnuhi\Exceptions\ConfigurationException;
 use PHPUnuhi\Models\Translation\TranslationSet;
@@ -73,6 +74,7 @@ class StorageFactory
         $this->storages[] = new PoStorage();
         $this->storages[] = new YamlStorage();
         $this->storages[] = new ResxStorage();
+        $this->storages[] = new StringsStorage();
         $this->storages[] = new Shopware6Storage();
     }